aarogyakosha/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
httpx==0.26.0
boto3==1.34.34
minio==7.2.3
pypdf2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
transformers==4.38.0
torch==2.2.1
sentence-transformers==2.3.1
meilisearch==0.28.0
celery==5.3.6
structlog==24.1.0
email-validator==2.1.0.post1
python-dateutil==2.8.2
aiofiles==23.2.1
python-magic==0.4.27
pydantic[email]==2.6.1
itsdangerous==2.1.2
